polished version of refactored config
