+++ lua/mappings/basic.lua
@@ -44,9 +44,3 @@ vim.api.nvim_set_keymap('n', '<A-Down>', ':resize -2<CR>', { noremap = true, sil
 vim.api.nvim_set_keymap('n', '<A-Left>', ':vertical resize +2<CR>', { noremap = true, silent = true })
 vim.api.nvim_set_keymap('n', '<A-Right>', ':vertical resize -2<CR>', { noremap = true, silent = true })
 
-
-vim.keymap.set("n", "<C-m>", function ()
-    local bufnr = vim.api.nvim_buf_get_number(0)
-    local cursor = vim.api.nvim_win_get_cursor(0)
-    require('checkbox_toggle').toggle_line(bufnr, cursor)
-end, { noremap = true, silent = true})
