+++ lua/plugins/render-markdown.lua
@@ -17,7 +17,7 @@ return {
             },
             custom = {
                 todo  = { raw = "[-]", rendered = "󰥔 ", highlight = "RenderMarkdownTodo", scope_highlight = nil },
-                warn  = { raw = "[/]", rendered = " ", highlight = "RenderMarkdownWarn", scope_highlight = nil },
+                warn  = { raw = "[/]", rendered = " ", highlight = "RenderMarkdownWarn", scope_highlight = "RenderMarkdownWarn" },
                 error = { raw = "[~]", rendered = "󰜺 ", highlight = "RenderMarkdownError", scope_highlight = "RenderMarkdownError" },
                 abort = { raw = "[_]", rendered = "󰚃 ", highlight = "RenderMarkdownHtmlComment", scope_highlight = "@markup.strikethrough" }
             }
